ata: move function definitions into a separate hpp file
Template instantiation with `extern template class` only works if the
functions are not defined inline. As we need to separate the definition
from the class definition anyway, move the function definitions into a
separate file, to have a concise header file for the interface.
